+++ src/modules/elements.js
@@ -5,5 +5,9 @@ const addList = document.querySelector('input[name=addtext]');
 const clear = document.getElementById('clear');
 
 export {
-  listBody, checkbox, taskDescription, addList, clear,
+    listBody,
+    checkbox,
+    taskDescription,
+    addList,
+    clear,
 };
